split into controlled and uncontrolled
