+++ src/js/components/Select/__tests__/SelectUncontrolled-test.js
@@ -13,8 +13,8 @@ import { Grommet } from '../..';
 import { Select } from '..';
 
 describe('Select', () => {
+  window.scrollTo = jest.fn();
   beforeEach(createPortal);
-
   afterEach(cleanup);
 
   test('should not have accessibility violations', async () => {
@@ -41,21 +41,8 @@ describe('Select', () => {
     expect(component.toJSON()).toMatchSnapshot();
   });
 
-  test('0 value', () => {
-    const component = renderer.create(
-      <Select
-        id="test-select"
-        placeholder="test select"
-        options={[0, 1]}
-        value={0}
-      />,
-    );
-    expect(component.toJSON()).toMatchSnapshot();
-  });
-
   test('opens and closes', () => {
     jest.useFakeTimers();
-    window.scrollTo = jest.fn();
     const onOpen = jest.fn();
     const onClose = jest.fn();
     const { getByPlaceholderText, container } = render(
@@ -111,38 +98,7 @@ describe('Select', () => {
     expectPortal('test-select__drop').toMatchSnapshot();
   });
 
-  test('search', () => {
-    jest.useFakeTimers();
-    const onSearch = jest.fn();
-    const { getByPlaceholderText, container } = render(
-      <Select
-        id="test-select"
-        placeholder="test select"
-        options={['one', 'two']}
-        onSearch={onSearch}
-        // need value here to show how options/value changes on search
-        value="two"
-      />,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-
-    fireEvent.click(getByPlaceholderText('test select'));
-
-    // advance timers so select can open
-    act(() => {
-      jest.advanceTimersByTime(200);
-    });
-    // snapshot on search box
-    expectPortal('test-select__drop').toMatchSnapshot();
-    expect(document.activeElement).toMatchSnapshot();
-    // add content to search box
-    fireEvent.change(document.activeElement, { target: { value: 'o' } });
-    expect(document.activeElement).toMatchSnapshot();
-    expect(onSearch).toBeCalledWith('o');
-  });
-
   test('select an option', () => {
-    window.scrollTo = jest.fn();
     const onChange = jest.fn();
     const { getByPlaceholderText, container } = render(
       <Select
@@ -167,36 +123,7 @@ describe('Select', () => {
     expect(window.scrollTo).toBeCalled();
   });
 
-  test('select an option with complex options', () => {
-    window.scrollTo = jest.fn();
-    const onChange = jest.fn();
-    const { getByText, container } = render(
-      <Select
-        id="test-select"
-        plain
-        value={<span>one</span>}
-        options={[{ test: 'one' }, { test: 'two' }]}
-        onChange={onChange}
-      >
-        {option => <span>{option.test}</span>}
-      </Select>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-
-    fireEvent.click(getByText('one'));
-
-    // pressing enter here nothing will happen
-    fireEvent.click(
-      document.getElementById('test-select__drop').querySelector('button'),
-    );
-    expect(onChange).toBeCalledWith(
-      expect.objectContaining({ value: { test: 'one' } }),
-    );
-    expect(window.scrollTo).toBeCalled();
-  });
-
   test('select an option with enter', () => {
-    window.scrollTo = jest.fn();
     const onChange = jest.fn();
     const { getByPlaceholderText, container } = render(
       <Select
@@ -229,100 +156,6 @@ describe('Select', () => {
     expect(window.scrollTo).toBeCalled();
   });
 
-  test('size', () => {
-    const component = renderer.create(
-      <Select
-        id="test-select"
-        size="large"
-        options={['one', 'two']}
-        selected={[]}
-        value={[]}
-        onChange={() => {}}
-      />,
-    );
-    expect(component.toJSON()).toMatchSnapshot();
-  });
-
-  test('multiple', () => {
-    const component = renderer.create(
-      <Select
-        id="test-select"
-        multiple
-        options={['one', 'two']}
-        selected={[]}
-        value={[]}
-      />,
-    );
-    expect(component.toJSON()).toMatchSnapshot();
-  });
-
-  test('multiple values', () => {
-    const { getByPlaceholderText, container } = render(
-      <Select
-        id="test-select"
-        placeholder="test select"
-        multiple
-        options={['one', 'two']}
-        selected={[0, 1]}
-        value={['one', 'two']}
-      />,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-
-    fireEvent.click(getByPlaceholderText('test select'));
-
-    expect(container.firstChild).toMatchSnapshot();
-    expectPortal('test-select__drop').toMatchSnapshot();
-  });
-
-  test('select another option', () => {
-    const onChange = jest.fn();
-    const { getByPlaceholderText, container } = render(
-      <Select
-        id="test-select"
-        placeholder="test select"
-        multiple
-        options={['one', 'two']}
-        onChange={onChange}
-        value={['two']}
-        selected={[1]}
-      />,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-
-    fireEvent.click(getByPlaceholderText('test select'));
-
-    fireEvent.click(
-      document.getElementById('test-select__drop').querySelector('button'),
-    );
-    expect(onChange).toBeCalledWith(
-      expect.objectContaining({ value: ['two', 'one'] }),
-    );
-  });
-
-  test('deselect an option', () => {
-    const onChange = jest.fn();
-    const { getByPlaceholderText, container } = render(
-      <Select
-        id="test-select"
-        placeholder="test select"
-        multiple
-        options={['one', 'two']}
-        onChange={onChange}
-        value={['one']}
-        selected={[0]}
-      />,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-
-    fireEvent.click(getByPlaceholderText('test select'));
-
-    fireEvent.click(
-      document.getElementById('test-select__drop').querySelector('button'),
-    );
-    expect(onChange).toBeCalledWith(expect.objectContaining({ value: [] }));
-  });
-
   test('disabled', () => {
     const { getByPlaceholderText, container } = render(
       <Select
@@ -340,25 +173,6 @@ describe('Select', () => {
     expect(container.firstChild).toMatchSnapshot();
     // dropdown should still be null because select is disabled
     expect(document.getElementById('test-select__drop')).toBeNull();
-  });
-
-  ['small', 'medium', 'large'].forEach(dropHeight => {
-    test(`${dropHeight} drop container height`, () => {
-      const { getByPlaceholderText } = render(
-        <Select
-          id="test-select"
-          size="large"
-          options={['one', 'two']}
-          selected={[]}
-          value={[]}
-          onChange={() => {}}
-          dropHeight={dropHeight}
-          placeholder="test select"
-        />,
-      );
-      fireEvent.click(getByPlaceholderText('test select'));
-      expect(document.activeElement).toMatchSnapshot();
-    });
   });
 
   test('empty results search', () => {
@@ -633,295 +447,7 @@ describe('Select', () => {
     expect(container.firstChild).toMatchSnapshot();
   });
 
-  test('onChange without valueKey', () => {
-    const onChange = jest.fn();
-    const Test = () => {
-      const [value] = React.useState();
-      return (
-        <Select
-          id="test-select"
-          placeholder="test select"
-          labelKey="name"
-          value={value}
-          options={[
-            {
-              id: 1,
-              name: 'Value1',
-            },
-            {
-              id: 2,
-              name: 'Value2',
-            },
-          ]}
-          onChange={onChange}
-        />
-      );
-    };
-    const { getByPlaceholderText, getByText, container } = render(
-      <Grommet>
-        <Test />
-      </Grommet>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-    fireEvent.click(getByPlaceholderText('test select'));
-
-    expectPortal('test-select__drop').toMatchSnapshot();
-
-    fireEvent.click(getByText('Value1'));
-    expect(onChange).toBeCalledWith(
-      expect.objectContaining({
-        value: {
-          id: 1,
-          name: 'Value1',
-        },
-      }),
-    );
-  });
-
-  test('multiple onChange without valueKey', () => {
-    const onChange = jest.fn();
-    const Test = () => {
-      const [value] = React.useState();
-      return (
-        <Select
-          id="test-select"
-          placeholder="test select"
-          labelKey="name"
-          value={value}
-          multiple
-          closeOnChange={false}
-          options={[
-            {
-              id: 1,
-              name: 'Value1',
-            },
-            {
-              id: 2,
-              name: 'Value2',
-            },
-          ]}
-          onChange={onChange}
-        />
-      );
-    };
-    const { getByPlaceholderText, getByText, container } = render(
-      <Grommet>
-        <Test />
-      </Grommet>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-    fireEvent.click(getByPlaceholderText('test select'));
-
-    expectPortal('test-select__drop').toMatchSnapshot();
-
-    fireEvent.click(getByText('Value1'));
-    expect(onChange).toBeCalledWith(
-      expect.objectContaining({
-        value: [
-          {
-            id: 1,
-            name: 'Value1',
-          },
-        ],
-      }),
-    );
-
-    expectPortal('test-select__drop').toMatchSnapshot();
-
-    fireEvent.click(getByText('Value2'));
-    expect(onChange).toBeCalledWith(
-      expect.objectContaining({
-        value: [
-          {
-            id: 1,
-            name: 'Value1',
-          },
-          {
-            id: 2,
-            name: 'Value2',
-          },
-        ],
-      }),
-    );
-  });
-
-  test('onChange with valueKey string', () => {
-    const onChange = jest.fn();
-    const Test = () => {
-      const [value] = React.useState();
-      return (
-        <Select
-          id="test-select"
-          placeholder="test select"
-          labelKey="name"
-          valueKey="id"
-          value={value}
-          options={[
-            {
-              id: 1,
-              name: 'Value1',
-            },
-            {
-              id: 2,
-              name: 'Value2',
-            },
-          ]}
-          onChange={onChange}
-        />
-      );
-    };
-    const { getByPlaceholderText, getByText, container } = render(
-      <Grommet>
-        <Test />
-      </Grommet>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-    fireEvent.click(getByPlaceholderText('test select'));
-
-    expectPortal('test-select__drop').toMatchSnapshot();
-
-    fireEvent.click(getByText('Value1'));
-    expect(onChange).toBeCalledWith(
-      expect.objectContaining({
-        value: {
-          id: 1,
-          name: 'Value1',
-        },
-      }),
-    );
-  });
-
-  test('multiple onChange with valueKey string', () => {
-    const onChange = jest.fn();
-    const Test = () => {
-      const [value] = React.useState();
-      return (
-        <Select
-          id="test-select"
-          placeholder="test select"
-          labelKey="name"
-          valueKey="id"
-          value={value}
-          multiple
-          options={[
-            {
-              id: 1,
-              name: 'Value1',
-            },
-            {
-              id: 2,
-              name: 'Value2',
-            },
-          ]}
-          onChange={onChange}
-        />
-      );
-    };
-    const { getByPlaceholderText, getByText, container } = render(
-      <Grommet>
-        <Test />
-      </Grommet>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-    fireEvent.click(getByPlaceholderText('test select'));
-
-    expectPortal('test-select__drop').toMatchSnapshot();
-
-    fireEvent.click(getByText('Value1'));
-    expect(onChange).toBeCalledWith(
-      expect.objectContaining({
-        value: [
-          {
-            id: 1,
-            name: 'Value1',
-          },
-        ],
-      }),
-    );
-  });
-
-  test('multiple onChange with valueKey reduce', () => {
-    const onChange = jest.fn();
-    const Test = () => {
-      const [value] = React.useState();
-      return (
-        <Select
-          id="test-select"
-          placeholder="test select"
-          labelKey="name"
-          valueKey={{ key: 'id', reduce: true }}
-          value={value}
-          multiple
-          options={[
-            {
-              id: 1,
-              name: 'Value1',
-            },
-            {
-              id: 2,
-              name: 'Value2',
-            },
-          ]}
-          onChange={onChange}
-        />
-      );
-    };
-    const { getByPlaceholderText, getByText, container } = render(
-      <Grommet>
-        <Test />
-      </Grommet>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-    fireEvent.click(getByPlaceholderText('test select'));
-
-    expectPortal('test-select__drop').toMatchSnapshot();
-
-    fireEvent.click(getByText('Value1'));
-    expect(onChange).toBeCalledWith(expect.objectContaining({ value: [1] }));
-  });
-
-  test('multiple onChange toggle with valueKey reduce', () => {
-    const onChange = jest.fn();
-    const Test = () => {
-      const [value] = React.useState([1]);
-      return (
-        <Select
-          id="test-select"
-          placeholder="test select"
-          labelKey="name"
-          valueKey={{ key: 'id', reduce: true }}
-          value={value}
-          multiple
-          options={[
-            {
-              id: 1,
-              name: 'Value1',
-            },
-            {
-              id: 2,
-              name: 'Value2',
-            },
-          ]}
-          onChange={onChange}
-        />
-      );
-    };
-    const { getByPlaceholderText, getByText, container } = render(
-      <Grommet>
-        <Test />
-      </Grommet>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-    fireEvent.click(getByPlaceholderText('test select'));
-
-    expectPortal('test-select__drop').toMatchSnapshot();
-
-    fireEvent.click(getByText('Value1'));
-    expect(onChange).toBeCalledWith(expect.objectContaining({ value: [] }));
-  });
-
   test('select an option then select a different option', () => {
-    window.scrollTo = jest.fn();
     const onChange = jest.fn();
     const { getByPlaceholderText } = render(
       <Select
@@ -1117,59 +643,5 @@ describe('Select', () => {
     fireEvent.click(getByPlaceholderText('test select'));
     expectPortal('test-select__drop').toMatchSnapshot();
   });
-
-  test('multiple with empty results', () => {
-    const { getByPlaceholderText, container } = render(
-      <Grommet>
-        <Select
-          id="test-select"
-          placeholder="test select"
-          options={['one', 'two']}
-          multiple
-          value={[]}
-        />
-      </Grommet>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-    fireEvent.click(getByPlaceholderText('test select'));
-    expectPortal('test-select__drop').toMatchSnapshot();
-  });
-
-  test('disabled key', () => {
-    jest.useFakeTimers();
-    const Test = () => {
-      const [value] = React.useState();
-      return (
-        <Select
-          id="test-select"
-          placeholder="test select"
-          labelKey="name"
-          valueKey="id"
-          value={value}
-          disabledKey="disabled"
-          options={[
-            {
-              id: 1,
-              name: 'Value1',
-              disabled: true,
-            },
-            {
-              id: 2,
-              name: 'Value2',
-              disabled: false,
-            },
-          ]}
-        />
-      );
-    };
-    const { getByPlaceholderText, container } = render(
-      <Grommet>
-        <Test />
-      </Grommet>,
-    );
-    expect(container.firstChild).toMatchSnapshot();
-    fireEvent.click(getByPlaceholderText('test select'));
-    jest.advanceTimersByTime(200);
-    expectPortal('test-select__drop').toMatchSnapshot();
-  });
+  window.scrollTo.mockRestore();
 });
